Use .laz file instead of .laz
