+++ modules/io/src/index.ts
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export {readLaz} from './addon';
+export {readLas} from './addon';
