+++ modules/io/src/rapidsai_io.ts
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export declare function readLaz(input: string): void;
+export declare function readLas(input: string): void;
